// contracts/generated/typescript/models/ProposalPreview.ts
export interface ProposalPreview {
    /**
     * 
     * @type {string}
     * @memberof ProposalPreview
     */
    proposalId: string;
    /**
     * 
     * @type {string}
     * @memberof ProposalPreview
     */
    targetType: string;
    /**
     * 
     * @type {ProposalPreviewRiskLevelEnum}
     * @memberof ProposalPreview
     */
    riskLevel: ProposalPreviewRiskLevelEnum;
    /**
     * 
     * @type {boolean}
     * @memberof ProposalPreview
     */
    stale: boolean;
    /**
     * 
     * @type {PreviewDocument}
     * @memberof ProposalPreview
     */
    base: PreviewDocument;
    /**
     * 
     * @type {PreviewDocument}
     * @memberof ProposalPreview
     */
    current: PreviewDocument;
    /**
     * 
     * @type {PreviewDocument}
     * @memberof ProposalPreview
     */
    proposed: PreviewDocument;
    /**
     * 
     * @type {DocumentDiff}
     * @memberof ProposalPreview
     */
    baseToCurrent: DocumentDiff;
    /**
     * 
     * @type {DocumentDiff}
     * @memberof ProposalPreview
     */
    baseToProposed: DocumentDiff;
    /**
     * 
     * @type {Array<OperationEvidence>}
     * @memberof ProposalPreview
     */
    evidence: Array<OperationEvidence>;
    /**
     * 
     * @type {PreviewImpact}
     * @memberof ProposalPreview
     */
    impact: PreviewImpact;
}


/**
 * @export
 */
export const ProposalPreviewRiskLevelEnum = {
    Low: 'low',
    Medium: 'medium',
    High: 'high',
    Critical: 'critical'
} as const;
export type ProposalPreviewRiskLevelEnum = typeof ProposalPreviewRiskLevelEnum[keyof typeof ProposalPreviewRiskLevelEnum];